% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Story_gen.r
\docType{data}
\name{character}
\alias{character}
\title{Story generator based on inputs}
\format{An object of class \code{character} of length 1.}
\usage{
character
}
\description{
This function generates stories.
}
\examples{
story_generator(character, place, ending)
story_gen()
Kadavresky()
}
\keyword{datasets}

